fix(gcp provider): move generate_client for consistency
